+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   ```
 - refactor: Expose `TooltipState`
 - fix: `NumberBox` calls `onChanged()` when clicked on ([#1196](https://github.com/bdlukaa/fluent_ui/issues/1196))
+- feat: Add `NumberBox.parser`, which allows to use a custom parse function for the number box. ([#1166](https://github.com/bdlukaa/fluent_ui/issues/1166))
 
 ## 4.13.0